Extract instance vertex types to quad.rs (refs #88)
Moves CellBgInstance, CellFgInstance, ImageQuadInstance, and
ensure_instance_buffer out of pipeline.rs into a new quad.rs module.
Mirrors wezterm-gui's quad.rs, which keeps per-instance vertex type
definitions separate from pipeline construction.

pipeline.rs now focuses solely on render pipeline setup (device/layout/
shader). quad.rs owns the public vertex/instance ABI consumed by the
emission loop and uniform layout.

No behavior change — pure mechanical refactor.
