+++ crates/amux-render-gpu/src/pipeline.rs
@@ -1,17 +1,7 @@
 use bytemuck::{Pod, Zeroable};
 use wgpu::util::DeviceExt;
 
-/// Per-cell background instance data.
-#[repr(C)]
-#[derive(Copy, Clone, Pod, Zeroable)]
-pub struct CellBgInstance {
-    /// Cell position in physical pixels (top-left corner).
-    pub pos: [f32; 2],
-    /// Cell size in physical pixels.
-    pub size: [f32; 2],
-    /// Background color (sRGB or linear RGBA, depending on render target format).
-    pub color: [f32; 4],
-}
+use crate::quad::{CellBgInstance, CellFgInstance, ImageQuadInstance};
 
 /// Viewport uniform data.
 #[repr(C)]
@@ -209,25 +199,6 @@ impl BackgroundPipeline {
 // ---------------------------------------------------------------------------
 // Foreground pipeline (textured glyph quads)
 // ---------------------------------------------------------------------------
-
-/// Per-glyph foreground instance data.
-#[repr(C)]
-#[derive(Copy, Clone, Pod, Zeroable)]
-pub struct CellFgInstance {
-    /// Glyph position in physical pixels (top-left corner).
-    pub pos: [f32; 2],
-    /// Glyph size in physical pixels.
-    pub size: [f32; 2],
-    /// Atlas UV min (top-left).
-    pub uv_min: [f32; 2],
-    /// Atlas UV max (bottom-right).
-    pub uv_max: [f32; 2],
-    /// Foreground color (sRGB or linear RGBA, depending on render target format).
-    pub color: [f32; 4],
-    /// 1.0 for color emoji, 0.0 for monochrome glyphs.
-    pub is_color: f32,
-    pub _pad: [f32; 3],
-}
 
 /// Foreground rendering pipeline: instanced textured quads for glyphs.
 ///
@@ -483,20 +454,6 @@ impl ForegroundPipeline {
 // Image pipeline (inline terminal images via Kitty protocol)
 // ---------------------------------------------------------------------------
 
-/// Per-quad image instance data.
-#[repr(C)]
-#[derive(Copy, Clone, Pod, Zeroable)]
-pub struct ImageQuadInstance {
-    /// Quad position in physical pixels (top-left corner).
-    pub pos: [f32; 2],
-    /// Quad size in physical pixels.
-    pub size: [f32; 2],
-    /// Texture UV min (top-left).
-    pub uv_min: [f32; 2],
-    /// Texture UV max (bottom-right).
-    pub uv_max: [f32; 2],
-}
-
 /// Image rendering pipeline: instanced textured quads for inline images.
 ///
 /// Each draw call binds a single image texture. Instance buffers provide
@@ -696,25 +653,4 @@ impl ImagePipeline {
         render_pass.set_index_buffer(self.index_buffer.slice(..), wgpu::IndexFormat::Uint16);
         render_pass.draw_indexed(0..6, 0, 0..instance_count);
     }
-}
-
-/// Create or grow an instance buffer if needed. Returns the buffer and its capacity.
-pub fn ensure_instance_buffer<T: Pod>(
-    device: &wgpu::Device,
-    existing: Option<&wgpu::Buffer>,
-    existing_capacity: usize,
-    required: usize,
-    label: &str,
-) -> Option<(wgpu::Buffer, usize)> {
-    if required <= existing_capacity && existing.is_some() {
-        return None; // existing buffer is fine
-    }
-    let new_capacity = required.max(1024).next_power_of_two();
-    let buffer = device.create_buffer(&wgpu::BufferDescriptor {
-        label: Some(label),
-        size: (new_capacity * std::mem::size_of::<T>()) as u64,
-        usage: wgpu::BufferUsages::VERTEX | wgpu::BufferUsages::COPY_DST,
-        mapped_at_creation: false,
-    });
-    Some((buffer, new_capacity))
 }
